+++ src/commands/stopMacro.ts
@@ -1,19 +1,25 @@
 import * as vscode from 'vscode';
 import { MacroRunInfo } from '../core/execution/macroRunInfo';
 import { Macro } from '../core/macro';
+import { StartupMacro } from '../core/startupMacro';
+import { getMacroUriFromStartupMacroUri } from '../core/startupMacroId';
 import { ExtensionContext } from '../extensionContext';
 
 export async function stopMacro(
   { log, runnerManager }: ExtensionContext,
-  uriOrMacroOrRunInfo: vscode.Uri | Macro | MacroRunInfo,
+  locator: vscode.Uri | Macro | MacroRunInfo | StartupMacro,
 ) {
-  const runInfos: MacroRunInfo[] =
-    uriOrMacroOrRunInfo instanceof Macro || uriOrMacroOrRunInfo instanceof vscode.Uri
-      ? [...runnerManager.getRunner(uriOrMacroOrRunInfo).runInstances]
-      : [uriOrMacroOrRunInfo];
+  const uri =
+    locator instanceof vscode.Uri
+      ? locator
+      : 'uri' in locator
+        ? getMacroUriFromStartupMacroUri(locator.uri)
+        : undefined;
 
-  for (const runInfo of runInfos) {
-    log.info('Requesting macro instance to stop via cancellation token', runInfo.runId);
-    runInfo.cts.cancel();
-  }
+  const runInfos: MacroRunInfo[] = uri
+    ? [...runnerManager.getRunner(uri).runInstances]
+    : [locator as MacroRunInfo];
+
+  log.info('Stopping macros via cancellation token —', ...runInfos.map(({ runId }) => runId));
+  runInfos.forEach(({ cts }) => cts.cancel());
 }