+++ src/core/library/libraryItem.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode';
 
 export type LibraryItemId = string;
 
-export interface LibraryItem {
-  id: LibraryItemId;
+export interface LibraryItem<TLibraryItemId extends LibraryItemId = LibraryItemId> {
+  id: TLibraryItemId;
   uri: vscode.Uri;
 }