+++ src/core/library/macroLibrary.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { isParent, UriLocator } from '../../utils/uri';
+import { isParent, resolveUri, UriLocator } from '../../utils/uri';
 import { isMacro, macroGlobPattern } from '../language';
 import { getMacroId } from '../macroId';
 import { Library } from './library';
@@ -43,15 +43,14 @@ export class MacroLibrary extends Library {
             .map((uri) => ({ id: getMacroId(uri), uri })),
         (_error) => [],
       );
-      this.initialized = true;
       this.addItems(...entries);
+      this.initialized = true;
     }
     const files = await super.getFiles();
     return files;
   }
 
   public override owns(locator: UriLocator): boolean {
-    const uri = locator instanceof vscode.Uri ? locator : locator.uri;
-    return isParent(this.uri, uri, { mustBeImmediate: true });
+    return isParent(this.uri, resolveUri(locator), { mustBeImmediate: true });
   }
 }